Drop Item and ItemPair types.
Seems a marker actually has a special value of type 'MARKER', so we can
represent this as a fairly conventional dict.

I am not certain whether ordering is important yet, so we'll stick with
SortableDict for now as that works pretty much the same and gives us
control over element ordering.