WC-3135: parser: Re-factor JSON handling
In 65fa7288240aab55d38af36e1551ad95ad6b6ad0 Phillipe made a change to
always return an array of grids, which is good from a consistency point
of view, but there's quite a bit of code out there that assumes just a
single grid.

In 5 years of dealing with Project Haystack, I've never actually seen a
response from a Haystack server with multiple "grids" in the one file
using ZINC (even though it supports it), and there's never officially
been a definition in JSON for a file with multiple grids.

So safe to assume, if someone hands a string in, it'll contain just one
grid, and they'll want just that grid parsed.  Add a flag `single`,
defaulting to `True`, which says "I just want the first grid you see".