WC-3135: datatypes: Always encode to UTF-8.
The existing code works if your locale uses a modern charset, but if
it's plain old `LANG=C` (i.e. like it is in a Docker container), it'll
fall flat because the charset there is `us-ascii`.

Really, two facts of life:
- we shouldn't be using ISO-8859-whatever, UTF-8 should be preferred.
- we shouldn't be relying on Python 2.7.

We can't do much about the second, but even Python 2.7 has good support
for Unicode these days.  We don't use ITA-2 anymore, and UTF-8 is
backward-compatible with ASCII anyways, so let's move on.